Add files via upload
The goal of this project was to create a Twitch.tv chat bot that would send a message into a Starcraft 2 Streamer's chat when it detected that they were supply blocked. accomplish this I used cv2 image thresholding, image scaling, image cropping, pytesseract to get an accurate value of the supply. Then I used tmi.js package and a little bit of Javascript file i/o
to only send a message when the txt file that the image detection program was writing to was updated.